Add jekyll build and local server as Grunt tasks
Previously we relied on `jekyll serve` (implicitly
`jekyll serve --watch`) to both do the jekyll build and
launch a development web server.

This substitutes Grunt tasks for those functions, using
`grunt-jekyll` and `grunt-contrib-connect -- rendering
the README incorrect for now. This improvement is
needed so that text transforms can be applied in the
correct order.

The default Grunt task now does:

1. clean -- Purge output directories
2. jekyll -- Does the jekyll build only (no serve, no watch), outputs to `_site`
3. replace -- Applies boldSyntaxElements to the generated `_site/index.html`
4. Copies needed files to `docs/`

A development server can then be launched with
`grunt connect`, which spawns a browser tab at `docs/'.
The `connect` task might be added at the end of the
default task, but it's not strictly necessary and
perhaps undesirable.

Also added `replace:preserveIndents`, but commented, because
it's strangely not finding matches. TODO