Adjust bitrate numbers
Previously, the bitrate for 5.0 was found to be a bit too restrictive
for common 4k content so had been adjusted slightly upward.
The bitrate for 5.1 had been adjusted in the same way to seem consistent.
HW implementers are happy with the increase for 5.0, but less for the 5.1
increase which causes problems.

This new change adjusts 5.0 upwards, and 5.1 downwards.
The change in this patch was approved on 5th June working group call.