Update 06.bitstream.syntax.md
In frame_size_with_refs( ), inside the for loop, if found_ref is equal to 1, we don't need to set the UpscaledWidth variable, because it will be immediately overwritten by the superres_params() call after the for loop.

Note: The current text is correct. This is just a cleanup. Please feel free to reject this.